Refactor data check method
